Refactoring main.py into util.py
