+++ pyblish/tests/test_main.py
@@ -1,5 +1,5 @@
 
-import pyblish.main
+import pyblish.util
 import pyblish.plugin
 
 from pyblish.vendor import mock
@@ -8,12 +8,12 @@ from pyblish.tests.lib import (
     teardown, FAMILY, HOST, setup_failing, setup_full)
 
 
-@mock.patch('pyblish.main.Publish.log')
+@mock.patch('pyblish.util.Publish.log')
 @with_setup(setup_full, teardown)
 def test_publish_all(_):
     """publish() calls upon each convenience function"""
     ctx = pyblish.plugin.Context()
-    pyblish.main.publish(context=ctx)
+    pyblish.util.publish(context=ctx)
 
     for inst in ctx:
         assert inst.data('selected') is True
@@ -22,18 +22,18 @@ def test_publish_all(_):
         assert inst.data('conformed') is True
 
 
-@mock.patch('pyblish.main.Publish.log')
+@mock.patch('pyblish.util.Publish.log')
 def test_publish_all_no_instances(mock_log):
     """Having no instances is fine, a warning is logged"""
     ctx = pyblish.plugin.Context()
-    pyblish.main.publish(ctx)
+    pyblish.util.publish(ctx)
     assert mock_log.warning.called
 
 
 @with_setup(setup_full, teardown)
 def test_publish_all_no_context():
     """Not passing a context is fine"""
-    ctx = pyblish.main.publish()
+    ctx = pyblish.util.publish()
 
     for inst in ctx:
         assert inst.data('selected') is True
@@ -42,12 +42,12 @@ def test_publish_all_no_context():
         assert inst.data('conformed') is True
 
 
-@mock.patch('pyblish.main.Publish.log')
+@mock.patch('pyblish.util.Publish.log')
 @with_setup(setup_full, teardown)
 def test_validate_all(_):
     """validate_all() calls upon two of the convenience functions"""
     ctx = pyblish.plugin.Context()
-    pyblish.main.validate_all(context=ctx)
+    pyblish.util.validate_all(context=ctx)
 
     for inst in ctx:
         assert inst.data('selected') is True
@@ -56,13 +56,13 @@ def test_validate_all(_):
         assert inst.data('conformed') is False
 
 
-@mock.patch('pyblish.main.Publish.log')
+@mock.patch('pyblish.util.Publish.log')
 @with_setup(setup_full, teardown)
 def test_convenience(_):
     """Convenience function work"""
     ctx = pyblish.plugin.Context()
 
-    pyblish.main.select(context=ctx)
+    pyblish.util.select(context=ctx)
 
     for inst in ctx:
         assert inst.data('selected') is True
@@ -70,7 +70,7 @@ def test_convenience(_):
         assert inst.data('extracted') is False
         assert inst.data('conformed') is False
 
-    pyblish.main.validate(context=ctx)
+    pyblish.util.validate(context=ctx)
 
     for inst in ctx:
         assert inst.data('selected') is True
@@ -78,7 +78,7 @@ def test_convenience(_):
         assert inst.data('extracted') is False
         assert inst.data('conformed') is False
 
-    pyblish.main.extract(context=ctx)
+    pyblish.util.extract(context=ctx)
 
     for inst in ctx:
         assert inst.data('selected') is True
@@ -86,7 +86,7 @@ def test_convenience(_):
         assert inst.data('extracted') is True
         assert inst.data('conformed') is False
 
-    pyblish.main.conform(context=ctx)
+    pyblish.util.conform(context=ctx)
 
     for inst in ctx:
         assert inst.data('selected') is True
@@ -95,17 +95,17 @@ def test_convenience(_):
         assert inst.data('conformed') is True
 
 
-@mock.patch('pyblish.main.Publish.log')
+@mock.patch('pyblish.util.Publish.log')
 @with_setup(setup_failing, teardown)
 def test_main_safe_processes_fail(_):
     """Failing selection, extraction or conform merely logs a message"""
     ctx = pyblish.plugin.Context()
-    pyblish.main.select(ctx)
+    pyblish.util.select(ctx)
 
     # Give plugins something to process
     inst = ctx.create_instance(name='TestInstance')
     inst.set_data('family', value=FAMILY)
     inst.set_data('host', value=HOST)
 
-    pyblish.main.extract(ctx)
-    pyblish.main.conform(ctx)
+    pyblish.util.extract(ctx)
+    pyblish.util.conform(ctx)
